+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 - add `Downloader::headers`
+- add `Downloader::modified`
 
 ## 0.2.0 - 2025-11-21
 - add `DownloaderBuilder::delay`